This PR fixes #6204 in two orthogonal ways:
- It solves an algorithmic issue in `shelve_unifiable` that made the computation of dependent evars cubic or quartic.
- It defines a fine-grained cache retaining the computation of free evars of a given hypothesis. This is similar in spirit to a previous tentative patch of @psteckler, except that the cache is not stored in the evar map (which can gobble up memory, see e.g. #4547) and that its granularity is at the level of hypotheses, not whole evars. This is done by checking for physical equality of the type of hypotheses. Doing it at the level of evars is not efficient enough to solve #6204.