@mattam82 2017-11-23
This looks great! Actually my assumption that we'd need something similar for typeclasses is wrong. The new proof search engine directly uses ``Proofview.unifiable`` so will benefit from it. It seems ``evar_dependencies`` and ``top_sort`` in class_tactics, that are used to order the set of goals coming from an evar_map (i.e. during type-checking) could also use the optimized ``filtered_undefined`` instead of ``undefined_evars_of..``.