@maximedenes 2017-11-23
> It seems evar_dependencies and top_sort in class_tactics, that are used to order the set of goals coming from an evar_map (i.e. during type-checking) could also use the optimized filtered_undefined instead of undefined_evars_of...

@mattam82 Can you open an issue for that so that we don't forget?